+++ src/daf/utils/__init__.py
@@ -1,0 +1,3 @@
+from daf._barrel import _public
+
+__all__ = _public()